requirements.txt
mpris_server
pyfoobeef
pyyaml
pygobject
